feature: extract feeds module and reindent React sources to 2 spaces per .editorconfig
