+++ react-app/src/App.css
@@ -1,44 +1,44 @@
 .layout {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
 }
 
 .layout-header {
-    display: flex;
-    align-items: center;
-    gap: 1.5rem;
-    padding: 0.75rem 1rem;
-    background: #ff6600;
-    color: #fff;
+  display: flex;
+  align-items: center;
+  gap: 1.5rem;
+  padding: 0.75rem 1rem;
+  background: #ff6600;
+  color: #fff;
 }
 
 .layout-logo {
-    font-weight: 700;
+  font-weight: 700;
 }
 
 .layout-nav {
-    display: flex;
-    gap: 1rem;
+  display: flex;
+  gap: 1rem;
 }
 
 .layout-nav a {
-    color: #fff;
-    text-decoration: none;
-    text-transform: capitalize;
+  color: #fff;
+  text-decoration: none;
+  text-transform: capitalize;
 }
 
 .layout-nav a.active {
-    text-decoration: underline;
+  text-decoration: underline;
 }
 
 .layout-main {
-    flex: 1;
-    padding: 1rem;
+  flex: 1;
+  padding: 1rem;
 }
 
 .layout-footer {
-    padding: 0.75rem 1rem;
-    font-size: 0.85rem;
-    border-top: 1px solid #ddd;
+  padding: 0.75rem 1rem;
+  font-size: 0.85rem;
+  border-top: 1px solid #ddd;
 }
