+++ react-app/src/App.tsx
@@ -5,5 +5,5 @@ import './App.css';
 const router = createBrowserRouter(routes);
 
 export default function App() {
-    return <RouterProvider router={router} />;
+  return <RouterProvider router={router} />;
 }
