+++ react-app/src/feeds.ts
@@ -1,0 +1,3 @@
+export const feeds = ['news', 'newest', 'show', 'ask', 'jobs'] as const;
+
+export type Feed = (typeof feeds)[number];
