+++ react-app/src/index.css
@@ -1,19 +1,19 @@
 :root {
-    font-family: system-ui, 'Segoe UI', Roboto, sans-serif;
-    line-height: 1.5;
-    color: #222;
-    background: #f6f6ef;
+  font-family: system-ui, 'Segoe UI', Roboto, sans-serif;
+  line-height: 1.5;
+  color: #222;
+  background: #f6f6ef;
 }
 
 * {
-    box-sizing: border-box;
+  box-sizing: border-box;
 }
 
 body {
-    margin: 0;
-    min-height: 100vh;
+  margin: 0;
+  min-height: 100vh;
 }
 
 a {
-    color: #ff6600;
+  color: #ff6600;
 }
