+++ react-app/src/main.tsx
@@ -4,7 +4,7 @@ import './index.css'
 import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
+ <StrictMode>
+  <App />
+ </StrictMode>,
 )
