+++ react-app/src/models/comment.ts
@@ -1,10 +1,10 @@
 export interface Comment {
-    id: number;
-    level: number;
-    user: string;
-    time: number;
-    time_ago: string;
-    content: string;
-    deleted: boolean;
-    comments: Comment[];
+  id: number;
+  level: number;
+  user: string;
+  time: number;
+  time_ago: string;
+  content: string;
+  deleted: boolean;
+  comments: Comment[];
 }
