+++ react-app/src/models/poll-result.ts
@@ -1,4 +1,4 @@
 export interface PollResult {
-    points: number;
-    content: string;
+  points: number;
+  content: string;
 }
