+++ react-app/src/models/settings.ts
@@ -1,7 +1,7 @@
 export interface Settings {
-    showSettings: boolean;
-    openLinkInNewTab: boolean;
-    theme: string;
-    titleFontSize: string;
-    listSpacing: string;
+  showSettings: boolean;
+  openLinkInNewTab: boolean;
+  theme: string;
+  titleFontSize: string;
+  listSpacing: string;
 }
