+++ react-app/src/models/user.ts
@@ -1,8 +1,8 @@
 export interface User {
-    id: string;
-    crated_time: number;
-    created: string;
-    karma: number;
-    avg: number;
-    about: string;
+  id: string;
+  crated_time: number;
+  created: string;
+  karma: number;
+  avg: number;
+  about: string;
 }
