+++ react-app/src/pages/FeedPage.tsx
@@ -1,18 +1,18 @@
 import { useParams } from 'react-router-dom';
-import type { Feed } from '../routes';
+import type { Feed } from '../feeds';
 
 interface FeedPageProps {
-    feedType: Feed;
+  feedType: Feed;
 }
 
 export default function FeedPage({ feedType }: FeedPageProps) {
-    const { page } = useParams<{ page: string }>();
+  const { page } = useParams<{ page: string }>();
 
-    return (
-        <section>
-            <h1>{feedType}</h1>
-            <p>Page {page}</p>
-            <p>Feed component not migrated yet.</p>
-        </section>
-    );
+  return (
+    <section>
+      <h1>{feedType}</h1>
+      <p>Page {page}</p>
+      <p>Feed component not migrated yet.</p>
+    </section>
+  );
 }
