+++ react-app/src/pages/ItemDetailsPage.tsx
@@ -1,12 +1,12 @@
 import { useParams } from 'react-router-dom';
 
 export default function ItemDetailsPage() {
-    const { id } = useParams<{ id: string }>();
+  const { id } = useParams<{ id: string }>();
 
-    return (
-        <section>
-            <h1>Item {id}</h1>
-            <p>Item details component not migrated yet.</p>
-        </section>
-    );
+  return (
+    <section>
+      <h1>Item {id}</h1>
+      <p>Item details component not migrated yet.</p>
+    </section>
+  );
 }
