+++ react-app/src/pages/NotFoundPage.tsx
@@ -1,10 +1,10 @@
 import { Link } from 'react-router-dom';
 
 export default function NotFoundPage() {
-    return (
-        <section>
-            <h1>Not found</h1>
-            <Link to="/news/1">Back to news</Link>
-        </section>
-    );
+  return (
+    <section>
+      <h1>Not found</h1>
+      <Link to="/news/1">Back to news</Link>
+    </section>
+  );
 }
