+++ react-app/src/pages/UserPage.tsx
@@ -1,12 +1,12 @@
 import { useParams } from 'react-router-dom';
 
 export default function UserPage() {
-    const { id } = useParams<{ id: string }>();
+  const { id } = useParams<{ id: string }>();
 
-    return (
-        <section>
-            <h1>User {id}</h1>
-            <p>User component not migrated yet.</p>
-        </section>
-    );
+  return (
+    <section>
+      <h1>User {id}</h1>
+      <p>User component not migrated yet.</p>
+    </section>
+  );
 }
